:sparkles: :: Company delete 기능 추가, Company Grades 평균 Scheduling 기능 추가, Company update 기능 추가

+++ Rolling-Core/src/main/java/com/stubee/rollingcore/domain/review/dto/response/ReviewInfoResponse.java
@@ -9,9 +9,9 @@ public record ReviewInfoResponse (
         String reviewPosition,
         String reviewCareerPath,
         Double totalGrade,
-        Short balanceGrade,
-        Short salaryGrade,
-        Short welfareGrade,
+        Double balanceGrade,
+        Double salaryGrade,
+        Double welfareGrade,
         LocalDateTime reviewCreatedAt,
         LocalDateTime reviewModifiedAt,
